Rename training_pt.py to training.py
