+++ sockeye/train.py
@@ -41,7 +41,7 @@ from . import loss_pt
 from . import lr_scheduler
 from . import model_pt
 from . import optimizers
-from . import training_pt
+from . import training
 from . import transformer_pt
 from . import utils
 from . import vocab
